backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
groq>=0.9.0
chromadb==0.5.3
python-dotenv==1.0.1
Pillow==10.3.0
pytest==8.2.2
httpx>=0.27.0
